refactor: modularize app orchestration flows into hooks
